Separate MotorcyclesController unit tests into different files. Add unit tests
