+++ TwoWheelsStable.Test/Controllers/Motorcycles/Update_Tests.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TwoWheelsStable.Test.Controllers.Motorcycles
+{
+    public class Update_Tests : MotorcyclesController_Tests_Base
+    {
+    }
+}
